[GraphBolt] refactor ondisk-related structures
