+++ python/dgl/graphbolt/impl/__init__.py
@@ -1,2 +1,4 @@
 """Implementation of GraphBolt."""
 from .ondisk_dataset import *
+from .ondisk_metadata import *
+from .torch_based_feature_store import *
