+++ tests/python/pytorch/graphbolt/test_feature_fetcher.py
@@ -6,10 +6,8 @@ import torch
 
 def get_graphbolt_fetch_func():
     feature_store = {
-        "feature": dgl.graphbolt.feature_store.TorchBasedFeatureStore(
-            torch.randn(200, 4)
-        ),
-        "label": dgl.graphbolt.feature_store.TorchBasedFeatureStore(
+        "feature": dgl.graphbolt.TorchBasedFeatureStore(torch.randn(200, 4)),
+        "label": dgl.graphbolt.TorchBasedFeatureStore(
             torch.randint(0, 10, (200,))
         ),
     }
